Quiet g++ compiler warnings Work around g++ compiler bug instantiating vector<NCVariable<Vector> > Added computeVariableExtents to (eventually) simplify data warehouses

+++ src/Uintah/Grid/UnionGeometryPiece.cc
@@ -24,14 +24,14 @@ UnionGeometryPiece::UnionGeometryPiece(const vector<GeometryPiece*>& child)
 
 UnionGeometryPiece::~UnionGeometryPiece()
 {
-  for (int i = 0; i < child.size(); i++) {
+  for (int i = 0; i < (int)child.size(); i++) {
     delete child[i];
   }
 }
 
 bool UnionGeometryPiece::inside(const Point &p) const 
 {
-  for (int i = 0; i < child.size(); i++) {
+  for (int i = 0; i < (int)child.size(); i++) {
     if (child[i]->inside(p))
       return true;
   }
@@ -48,7 +48,7 @@ Box UnionGeometryPiece::getBoundingBox() const
   lo = child[0]->getBoundingBox().lower();
   hi = child[0]->getBoundingBox().upper();
 
-  for (int i = 0; i < child.size(); i++) {
+  for (int i = 0; i < (int)child.size(); i++) {
     Box box = child[i]->getBoundingBox();
     lo = Min(lo,box.lower());
     hi = Max(hi,box.upper());
